+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- `ResultExt` is now a collection of Emacs-specific extension methods for `Result`, instead of a re-export of `failure::ResultExt`.
+- Added `failure` as a re-exported sub-module.
+- Added `expect_descriptive` for better error message when converting `Result` to panic.
 
 ## [0.8.0] - 2019-04-20
 - Input parameters with reference types are now interpreted as Rust data structures embedded in `user-ptr` objects.